+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 Yii2 multiple input change log
 ==============================
 
-1.2.6 in development
+1.2.7 in development
 --------------------
 
+1.2.6
+-----
+
 - Bug #49: urlencoded field token replacement in js template (rolmonk)
+- Enh #48: Added option `min` for setting minimum number of rows
+- Enh: added option 'addButtonPosition'
 
 1.2.5
 -----